Refactor components into their own files
